Update dict_de.po
Just filled some empty translations, and suggest some changes.

For Example the translation for "calculate" was "kalkulieren" that word does nobody use for this purpose. Its "berechnen".
And I think there could be some optimisation in string length. Like "Copy as Hex" "Als Hexadezimal kopieren" that's an uncommon long form. In german we just use hex. An Hex-File is an "Hexdatei".

For use in an software, some long phrases would mostly not fit right in the layout I think... Maybe it is useful for you.