Update dict_zh.po
Finished Chinese translation
PS:
I am not sure where the "TEXT" property code of "Controls" is. If it represents a control of a UI window, such as a Label, GroupBox, etc., then it should be translated into "控件". If it is the text above the configuration parameter option, then it should be translated into "控制".
I hope someone familiar with the software can correct or confirm it.