Fix typo due to unsorted file structure
I encountered issues with missing translations because the file wasn't sorted. I initially prepared the file in a format that allowed me to type entries like "load " and "load database," but due to the unsorted structure, I had to jump between multiple lines to make changes. This caused some entries to be missed.

I will fix the remaining errors and sort the file to prevent such mistakes in the future. I’ll start with this Swedish file and then, if possible, sort the other files. While sorting, I will also check for typos in languages I’m familiar with.

Thank you for your understanding.